updated constants
