+++ src/pages/AnalyticsCompare/AnalyticsChart/AnalyticsChart.tsx
@@ -10,9 +10,7 @@ import {
 import { useAdapter } from 'hooks';
 import { faChartBar } from 'icons/regular';
 import { activeNetworkSelector } from 'redux/selectors';
-
-import { ChartResolutionSelector } from './components/ChartResolution';
-import { ChartResolutionRangeType } from './components/ChartResolution/types';
+import { ChartResolutionRangeType } from 'types';
 
 import { getChartColorPalette } from '../helpers/getChartColorPalette';
