+++ src/types/index.ts
@@ -2,6 +2,7 @@ export * from './account.types';
 export * from './accountStake.types';
 export * from './adapter.types';
 export * from './block.types';
+export * from './charts.types';
 export * from './collection.types';
 export * from './economics.types';
 export * from './events.types';
